+++ CHANGELOG.md
@@ -18,5 +18,8 @@
 - Improving testability of functions
   Breaking the app into smaller files and adding tests
 ## 3.1.0
-- Added tests for func TestFileVersioner_WriteVersion(t *testing.T) {
-  // Create a temporary file
+- Added tests for func WriteVersion
+  Improving testability and adding tests
+## 3.2.0
+- Added more tests
+  Improved testability and added more test use cases